Staging: dgnc: dgnc_driver.c: Add end tag to dgnc_init_module()
The work flow of dgnc_init_module() is base on if statment. Here is a
nest "if" in the function.

To avoid the nest if statment, make the code easier to understand and
make the codingstyle to be identical with others, add "end" label to
dgnc_init_module() and use goto to control the work flow.

Signed-off-by: Navy Cheng <navych@126.com>